[wip][fix] remove "as usize" in cpu.reg index

+++ src/cpu/execution/exe_inst_16.rs
@@ -13,63 +13,63 @@ pub fn exe_cinst(inst: &Instruction, cpu: &mut CPU, dram: &mut Dram) {
 
     match inst.opc {
         OP_C_LI => {
-            cpu.reg[inst.rd.unwrap() as usize] = inst.imm.unwrap();
+            cpu.reg[inst.rd.unwrap()] = inst.imm.unwrap();
         },
         OP_C_LW => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                Dram::load32(dram, (inst.rs1.unwrap() as i32 + inst.imm.unwrap()) as usize);
+            cpu.reg[inst.rd.unwrap()] =
+                Dram::load32(dram, (inst.rs1.unwrap() as i32 + inst.imm.unwrap()));
         },
         OP_C_LWSP => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                Dram::load32(dram, (cpu.reg[REG_SP] as i32 + inst.imm.unwrap()) as usize);
+            cpu.reg[inst.rd.unwrap()] =
+                Dram::load32(dram, (cpu.reg[REG_SP] as i32 + inst.imm.unwrap()));
         },
         OP_C_LUI => {
-            cpu.reg[inst.rd.unwrap() as usize] = inst.imm.unwrap() << 12;
+            cpu.reg[inst.rd.unwrap()] = inst.imm.unwrap() << 12;
         },
         OP_C_SW => {},
         OP_C_SLLI => {},
         OP_C_SWSP => {},
         OP_C_SRLI => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] >> inst.imm.unwrap() as i32;
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] >> inst.imm.unwrap() as i32;
         },
         OP_C_SRAI => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                (cpu.reg[inst.rs1.unwrap() as usize] as i32) >> inst.imm.unwrap() as i32;
+            cpu.reg[inst.rd.unwrap()] =
+                (cpu.reg[inst.rs1.unwrap()] as i32) >> inst.imm.unwrap() as i32;
         },
         OP_C_ADD => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] + cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] + cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_ADDI4SPN => {
             cpu.reg[REG_SP] = 
-                cpu.reg[REG_SP] + ((cpu.reg[inst.imm.unwrap() as usize] >> 2) & 0x1FF);
+                cpu.reg[REG_SP] + ((cpu.reg[inst.imm.unwrap()] >> 2) & 0x1FF);
         },
         OP_C_ADDI => {
-            cpu.reg[inst.rd.unwrap() as usize] += inst.rs1.unwrap() as i32;
+            cpu.reg[inst.rd.unwrap()] += inst.rs1.unwrap() as i32;
         },
         OP_C_ADDI16SP => {
             cpu.reg[REG_SP] = 
-                cpu.reg[REG_SP] + ((cpu.reg[inst.imm.unwrap() as usize] >> 4) & 0x1FF);
+                cpu.reg[REG_SP] + ((cpu.reg[inst.imm.unwrap()] >> 4) & 0x1FF);
         },
         OP_C_ANDI => {
-            cpu.reg[inst.rd.unwrap() as usize] &= inst.rs1.unwrap() as i32;
+            cpu.reg[inst.rd.unwrap()] &= inst.rs1.unwrap() as i32;
         },
         OP_C_SUB => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] - cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] - cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_XOR => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] ^ cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] ^ cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_OR => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] | cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] | cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_AND => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] & cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs1.unwrap()] & cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_J => {
             cpu.pc += inst.imm.unwrap() as u32;
@@ -80,24 +80,24 @@ pub fn exe_cinst(inst: &Instruction, cpu: &mut CPU, dram: &mut Dram) {
         },
         OP_C_JALR => {
             cpu.reg[LINK_REG] = (cpu.pc + INST_SIZE) as i32; 
-            cpu.pc += (cpu.reg[inst.rs1.unwrap() as usize]  + inst.imm.unwrap()) as u32;
+            cpu.pc += (cpu.reg[inst.rs1.unwrap()]  + inst.imm.unwrap()) as u32;
         },
         OP_C_BEQZ => {
-            if cpu.reg[inst.rs1.unwrap() as usize] == 0 {
+            if cpu.reg[inst.rs1.unwrap()] == 0 {
                 cpu.pc += inst.imm.unwrap() as u32;
             } 
         },
         OP_C_BNEZ => {
-            if cpu.reg[inst.rs1.unwrap() as usize] != 0 {
+            if cpu.reg[inst.rs1.unwrap()] != 0 {
                 cpu.pc += inst.imm.unwrap() as u32;
             } 
         },
         OP_C_JR => {
-            cpu.pc += cpu.reg[inst.rs1.unwrap() as usize] as u32;
+            cpu.pc += cpu.reg[inst.rs1.unwrap()] as u32;
         },
         OP_C_MV => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs2.unwrap() as usize];
+            cpu.reg[inst.rd.unwrap()] =
+                cpu.reg[inst.rs2.unwrap()];
         },
         OP_C_EBREAK => {},
         OP_C_NOP => {/* NOP */},
